Added key bindings, fixed some bugs, changed some functions to C++11, optimized code, added UI size scaling depending on window size, moved all texture code back to Texture class, left just unique properties which control texture

+++ Space Shooter/Bullet.h
@@ -21,7 +21,7 @@ namespace Engine
 	class Bullet : public BaseGameObject
 	{
 		public:
-			Bullet(int, int, glm::vec2, glm::vec2, glm::vec4);
+			Bullet(float, float, glm::vec2, glm::vec2, glm::vec4);
 			bool update(float);
 			void onCollision(BaseGameObject*, BaseGameObject*);
 	};
